Update update-contact.ts
Problem: Mautic PATCH tries to write to Mautic it will put  //edit in the URL and one of those needs to be removed. I believe this edit in the code will fix the problem, but I could be wrong this is why I am trying to tell you the problem here. Here is what I believe my code edit will do: 

1. If the base_url ends with a slash, it removes that slash (using slice(0, -1)).

2. If the base_url doesn't end with a slash, it uses it as is.

3. Then it adds a single slash before appending the rest of the API endpoint.

ERROR LOG BELOW when trying to do a contact update using Mautic Active Pieces module (to be more specific I used the "Update Contact with contact ID and yes it pulls all the data correctly, but it cant save to contact because of URL error). 

{
  "response": {
    "status": 404,
    "body": {
      "errors": [
        {
          "message": "Requested URL not found: /api/contacts//edit",
          "code": 404,
          "type": null
        }